Add charts to Workspace (#20)
Description: added a pop-up window that appears after selecting a point on the map. The window contains charts displaying data for the selected point and sliders for choosing the year range. The window minimizes when the main window becomes inactive and gets restored when the main window becomes active.

Note the following:

the window isn't actually minimized, it's just moved out of the screen
when the window is restored, it's resized and repositioned to the original values because to constantly keep track of its dimensions and position requires knowing when the user manually minimizes it and that's difficult and adds a lot of complexity